+++ src/main/java/fuzs/configmenusforge/ConfigMenusForge.java
@@ -31,12 +31,11 @@ import java.io.File;
 public class ConfigMenusForge {
 
     public static final String MOD_ID = "configmenusforge";
-    public static final String NAME = "Config Menus for Forge";
-    public static final String URL = "https://www.curseforge.com/minecraft/mc-mods/config-menus-forge";
-    public static final Logger LOGGER = LogManager.getLogger(ConfigMenusForge.NAME);
-
-    @SuppressWarnings("Convert2MethodRef")
-    public static final IProxy PROXY = EnvTypeExecutor.runForDist(() -> () -> new ClientProxy(), () -> () -> new ServerProxy());
+    public static final String MOD_NAME = "Config Menus for Forge";
+    public static final String MOD_URL = "https://www.curseforge.com/minecraft/mc-mods/config-menus-forge";
+    public static final Logger LOGGER = LogManager.getLogger(ConfigMenusForge.MOD_NAME);
+    
+    public static final NetworkHandler NETWORK = NetworkHandler.of(MOD_ID, true, false);
 
     @SubscribeEvent
     public static void onConstructMod(final FMLConstructModEvent evt) {
@@ -46,10 +45,10 @@ public class ConfigMenusForge {
     }
 
     private static void registerMessages() {
-        NetworkHandler.INSTANCE.register(C2SAskPermissionsMessage.class, C2SAskPermissionsMessage::new, NetworkDirection.PLAY_TO_SERVER);
-        NetworkHandler.INSTANCE.register(S2CGrantPermissionsMessage.class, S2CGrantPermissionsMessage::new, NetworkDirection.PLAY_TO_CLIENT);
-        NetworkHandler.INSTANCE.register(C2SSendConfigMessage.class, C2SSendConfigMessage::new, NetworkDirection.PLAY_TO_SERVER);
-        NetworkHandler.INSTANCE.register(S2CUpdateConfigMessage.class, S2CUpdateConfigMessage::new, NetworkDirection.PLAY_TO_CLIENT);
+        NETWORK.register(C2SAskPermissionsMessage.class, C2SAskPermissionsMessage::new, NetworkDirection.PLAY_TO_SERVER);
+        NETWORK.register(S2CGrantPermissionsMessage.class, S2CGrantPermissionsMessage::new, NetworkDirection.PLAY_TO_CLIENT);
+        NETWORK.register(C2SSendConfigMessage.class, C2SSendConfigMessage::new, NetworkDirection.PLAY_TO_SERVER);
+        NETWORK.register(S2CUpdateConfigMessage.class, S2CUpdateConfigMessage::new, NetworkDirection.PLAY_TO_CLIENT);
     }
 
     private static void addTestConfigs() {