+++ src/main/java/fuzs/configmenusforge/lib/proxy/IProxy.java
@@ -1,5 +1,6 @@
 package fuzs.configmenusforge.lib.proxy;
 
+import fuzs.configmenusforge.lib.core.EnvTypeExecutor;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.server.MinecraftServer;
 
@@ -7,6 +8,11 @@ import net.minecraft.server.MinecraftServer;
  * proxy base class
  */
 public interface IProxy {
+    /**
+     * sided proxy depending on physical side
+     */
+    @SuppressWarnings("Convert2MethodRef")
+    IProxy INSTANCE = EnvTypeExecutor.runForDist(() -> () -> new ClientProxy(), () -> () -> new ServerProxy());
 
     /**
      * @return client player from Minecraft singleton when on physical client, otherwise null